# dist/openapi.yaml
...
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API describes recipes site.
  contact:
    email: carmelit@post.bgu.ac.il
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: "http://127.0.0.1:3000"
  description: SwaggerHub API for recipes site
# tags:
# - name: admins
#   description: Secured Admin-only calls
# - name: developers
#   description: Operations available to regular developers
paths:
  /recipes/{id}:
    get:
      tags:
      - recipes
      summary: Get a recipe by ID
      operationId: getRecipeById
      parameters:
      - name: id
        in: path
        description: ID of the recipe to get
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: recipe found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        "404":
          description: Recipe not found
  /users/addRecipe/:
    post:
      tags:
      - user
      summary: add new recipe
      description: Adds new recipe to the system
      operationId: addRecipe
      requestBody:
        description: new recipe to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewRecipe'
      responses:
        "201":
          description: recipe created
        "400":
          description: "Parameters Error"
  /users/myRecipes/{id}:
    get:
      tags:
      - user
      summary: get recipe created by the user
      description: get full recipe created by the user
      operationId: myRecipes
      parameters:
      - name: id
        in: path
        description: ID of the recipe to get
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "201":
          description: recipe created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyRecipe'
        "401":
          description: "Unauthorized"
  /users/myRecipes/:
    get:
      tags:
      - user
      summary: get all recipes previews created by the user
      description: get previews of recipe created by the user
      operationId: myRecipesPreviews
      responses:
        "200":
          description: all previews retrived
          content:
            application/json:
              schema:
              type: array
              items:
                $ref: '#/components/schemas/MyRecipeDetails'
        "401":
          description: "Unauthorized"
        "404":
        description: "Wrong type of recipe."
  /users/myRecipes/family/{id}:
    get:
      tags:
      - user
      summary: get recipe created by the user (recipes from type family)
      description: get full recipe created by the user (recipes from type family)
      operationId: myFamilyRecipes
      parameters:
      - name: id
        in: path
        description: ID of the recipe to get (FAXXX)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "201":
          description: recipe created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyRecipe'
        "401":
          description: "Unauthorized"
  /users/myRecipes/family:
    get:
      tags:
      - user
      summary: get all recipes previews created by the user (recipes from type family)
      description: get previews of recipe created by the user (recipes from type family)
      operationId: myFamilyRecipesPreviews
      responses:
        "200":
          description: all previews retrived
          content:
            application/json:
              schema:
              type: array
              items:
                $ref: '#/components/schemas/MyRecipeDetails'
        "401":
          description: "Unauthorized"
        "404":
        description: "Wrong type of recipe."
  /recipes/preview/{id}:
    get:
      tags:
      - recipes
      summary: get recipe preview
      description: user get preview of recipe.
      operationId: recipe preview
      parameters:
      - name: id
        in: path
        description: ID of the recipe to get preview
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful preview of all recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeDetails'
        "404":
          description: recipes not found
...
  /users/lastViews:
    get:
      tags:
      - user
      summary: get last 3 recipes watched by the user - (db lastviews).
      description: display preview of last 3 watched recipes.
      operationId: lastViewsUpdate
      responses:
        "200":
          description: Previews retrived successfully.
          content:
            application/json:
              schema:
                type: array
                description: array of 3 previews of reccipa
                items:
                - $ref: '#/components/schemas/RecipeDetails'
                - $ref: '#/components/schemas/MyRecipeDetails'
        "404":
          description: Server error.
...
  /users/favorites:
    post:
        tags:
        - user
        summary: user favorite recipes
        description: Save recipe in user favorite.
        operationId: addFavorite
        requestBody:
          description: recipeid
          content:
            application/json:
              schema:
                required:
                - recipeid
                properties:
                  recipeid:
                    type: int
        responses:
          "303":
            description: "User created successfully, redirecting to login page"
            headers:
              location:
                description: redirect to login page
                style: simple
                explode: false
                schema:
                  type: string
          "400":
            description: Invalid request body
    get:
        tags:
        - user
        summary: user favorite recipes
        description: get all recipes saved in user favorites.
        operationId: getFavorite, return all favorite recipes of user
        responses:
          "200":
            description: Recipes retrived succedsfully
            content:
            application/json:
              schema:
              type: object
              properties:
              - recipeid: 
                type: int
              - preview:
                $ref: '#/components/schemas/RecipeDetails'
          "400":
            description: Invalid request body
  /users/recipes/{id}:
    get:
      tags:
      - user
      summary: get prepare meal page for the user 
      description:  get prepare meal page for the user 
      operationId: prepareMeal
      parameters:
      - name: id
        in: path
        description: ID of the recipe to prepare
      responses:
        "200":
          description: prepare page created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreparedRecipe'
        "401":
          description: "Unauthorized"
        "404":
          description: "Page doesnt exist."
  /users/recipes:
    get:
      tags:
      - user
      summary: get all recipes user want tp prepare. 
      description:  get all recipes user want tp prepare. 
      operationId: allPrepareMeals
      responses:
        "200":
          description: all reciped desplayed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDetails'
        "401":
          description: "Unauthorized"
    post:
      tags:
      - user
      summary: add meal to "prepare meals"
      description:  add meal to "prepare meals"
      operationId: addToAllPrepareMeals
      requestBody:
        description: recipeid
        content:
          application/json:
            schema:
              required:
              - recipeid
              properties:
                recipeid:
                  type: string
        responses:
          "200":
            description: meal addedd successfully.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/RecipeDetails'
          "401":
            description: "Unauthorized"
          "404":
            description: "Recipe doesnt exist."
    delete:
      tags:
      - user
      summary: delete meal from "prepare meals"
      description:  delete meal from "prepare meals"
      operationId: delFromAllPrepareMeals
      requestBody:
        description: recipeid
        content:
          application/json:
            schema:
              required:
              - recipeid
              - userid
              properties:
                recipeid:
                  type: string
                userid:
                  type: integer
        responses:
          "200":
            description: meal deleted successfully.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/RecipeDetails'
          "401":
            description: "Unauthorized"
components:
  schemas:
...
    RecipeDetails:
      required:
      - id
      - title
      - readyInMinutes
      - image
      - popularity
      - vegan
      - vegetarian
      - glutenFree
      - favorite
      - watched
      type: object
      properties:
        id:
          type: integer
          example: 1
        image:
          type: string
          format: url
          example: "recipesImage\\{id=19}"
        title:
          type: string
          example: Chocolate Cake
        readyInMinutes:
          type: int
          format: duration
          example: 20
        popularity:
          type: integer
          example: 456
        vegan:
          type: boolean
          example: true
        vegetarian:
          type: boolean
          example: true
        glutenFree:
          type: boolean
          example: true
        favorite:
          type: boolean
          example: true
        watched:
          type: boolean
          example: true
    Ingredient:
      type: object
      required:
      - name
      - quantity
      - unit
      properties:
        name:
          type: string
          example: sugar
        quantity:
          type: number
          example: 1.25
        unit:
          type: string
          example: spoons
    Recipe:
      required:
      - ingredients
      - instructions
      - servings
      - recipe
      type: object
      properties:
        recipe:
          $ref: '#/components/schemas/RecipeDetails'
        ingredients:
          type: array
          example:
          - name: sugar
            quantity: 1.25
            unit: spoons
          - name: flour
            quantity: 2
            unit: cups
          - name: chocolate chips
            quantity: 0.5
            unit: cups
          items:
            $ref: '#/components/schemas/Ingredient'
        instructions:
          type: array
          example: ["Preheat the oven to 210 C.", "Mix the flour, sugar, and salt in a large bowl."]
        servings:
          type: integer
          example: 12
...
    MyRecipeDetails:
        required:
        - id
        - title
        - readyInMinutes
        - image
        - vegan
        - vegetarian
        - glutenFree
        type: object
        properties:
          id:
            type: integer
            example: 1
          image:
            type: string
            format: url
            example: "recipesImage\\{id=19}"
          title:
            type: string
            example: Chocolate Cake
          readyInMinutes:
            type: int
            format: duration
            example: 20
          vegan:
            type: boolean
            example: true
          vegetarian:
            type: boolean
            example: true
          glutenFree:
            type: boolean
            example: true
    MyRecipe:
      required:
      - ingredients
      - instructions
      - servings
      - recipe
      type: object
      properties:
        recipe:
          $ref: '#/components/schemas/MyRecipeDetails'
        ingredients:
          type: array
          example:
          - name: sugar
            quantity: 1.25
            unit: spoons
          - name: flour
            quantity: 2
            unit: cups
          - name: chocolate chips
            quantity: 0.5
            unit: cups
          items:
            $ref: '#/components/schemas/Ingredient'
        instructions:
          type: array
          example: ["Preheat the oven to 210 C.", "Mix the flour, sugar, and salt in a large bowl."]
        servings:
          type: integer
          example: 12
    NewRecipe:
      required:
      - recipe
      - type
      - optionalDescription
      type: object
      properties:
        recipe:
          required: true
          $ref: '#/components/schemas/Recipe'
        type:
          required: true
          type: string
          example: Family\Private
        optionalDescription:
          required: false
          type: string
          example: Recipe made by grandmother. Prepared in Sukot holiday.
    PreparedRecipe:
      required:
      - equipment
      - ingredients
      - number
      - step
      - length
      type: object
      properties:
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/Equipment' 
        ingredients:
          type: array
          example:
          - name: sugar
        number:
          type: int
          example: 3
        step:
          type: string
          example: "Transfer the pancakes to a platter and keep warm in a 200 degree F oven."
    Equipment:
      reqiured:
      - name
      additional:
      - temperature
      properties:
        - name: number
          type: float
        - name: unit
          type: string
      example:
        - name: "oven"
          temperature: 
          number: 200.0
          unit: "Fahrenheit"